refactor: modularize AST and implement missing DML features
- Refactored monolithic ast.rs into a modular ast/ directory structure.
- Implemented TOP clause for UPDATE and DELETE statements.
- Added support for INSERT ... EXEC to insert stored procedure results.
- Implemented T-SQL table hints (WITH clause) in the parser.
- Fixed trigger pseudo-table resolution by normalizing schema handling.
- Improved multi-CTE scoping in ScriptExecutor to support sequential references.
- Updated documentation and crate exports to maintain public API compatibility.
