+++ crates/tsql_core/src/parser/statements/dml/insert.rs
@@ -4,7 +4,7 @@ use crate::parser::utils::{find_keyword_top_level, parse_object_name};
 use super::output::parse_output_clause;
 
 pub(crate) fn parse_insert(sql: &str) -> Result<Statement, DbError> {
-    let mut after_into = sql["INSERT INTO".len()..].trim();
+    let mut after_into = sql["INSERT".len()..].trim();
     if after_into.to_uppercase().starts_with("INTO ") {
         after_into = after_into[5..].trim();
     }
@@ -15,9 +15,7 @@ pub(crate) fn parse_insert(sql: &str) -> Result<Statement, DbError> {
         return Ok(Statement::Insert(InsertStmt {
             table: parse_object_name(table_name),
             columns: None,
-            values: vec![],
-            default_values: true,
-            select_source: None,
+            source: InsertSource::DefaultValues,
             output: None,
             output_into: None,
         }));
@@ -28,15 +26,13 @@ pub(crate) fn parse_insert(sql: &str) -> Result<Statement, DbError> {
 
     // Determine OUTPUT content
     let (output, output_into, after_into_no_output) = if let Some(oi) = output_idx {
-        // OUTPUT INSERTED.col1, DELETED.col1
-        // The OUTPUT is between table/columns and VALUES/SELECT
         let after_output = &after_into[oi + "OUTPUT".len()..];
         let end_idx = find_keyword_top_level(after_output, "VALUES")
             .or_else(|| find_keyword_top_level(after_output, "SELECT"))
+            .or_else(|| find_keyword_top_level(after_output, "EXEC"))
             .unwrap_or(after_output.len());
         let output_raw = &after_output[..end_idx];
         let (parsed_output, into_target) = parse_output_clause(output_raw.trim())?;
-        // Reconstruct: table part + VALUES/SELECT part (skipping OUTPUT columns)
         let before_output = &after_into[..oi];
         let after_output_cols = &after_output[end_idx..];
         let reconstructed = format!("{}{}", before_output, after_output_cols);
@@ -50,51 +46,66 @@ pub(crate) fn parse_insert(sql: &str) -> Result<Statement, DbError> {
     // Re-check with stripped string
     let select_idx_stripped = find_keyword_top_level(after_into_stripped, "SELECT");
     let values_idx_stripped = find_keyword_top_level(after_into_stripped, "VALUES");
+    let exec_idx_stripped = find_keyword_top_level(after_into_stripped, "EXEC");
 
     if let Some(sel_idx) = select_idx_stripped {
-        // INSERT ... SELECT
-        if values_idx_stripped.is_none() || sel_idx < values_idx_stripped.unwrap() {
+        // Check if SELECT is before VALUES and EXEC
+        if (values_idx_stripped.is_none() || sel_idx < values_idx_stripped.unwrap()) &&
+           (exec_idx_stripped.is_none() || sel_idx < exec_idx_stripped.unwrap()) {
             let head = after_into_stripped[..sel_idx].trim();
-            let (table_name, columns) = if let Some(open) = head.find('(') {
-                let close = head
-                    .rfind(')')
-                    .ok_or_else(|| DbError::Parse("missing ')' in INSERT columns".into()))?;
-                let table_name = head[..open].trim();
-                let cols = head[open + 1..close]
-                    .split(',')
-                    .map(|c| c.trim().to_string())
-                    .collect::<Vec<_>>();
-                (table_name.to_string(), Some(cols))
-            } else {
-                (head.to_string(), None)
-            };
-
+            let (table_name, columns) = parse_table_and_columns(head)?;
             let select_sql = &after_into_stripped[sel_idx..];
             let select_stmt = crate::parser::statements::select::parse_select(select_sql)?;
             if let Statement::Select(sel) = select_stmt {
                 return Ok(Statement::Insert(InsertStmt {
                     table: parse_object_name(&table_name),
                     columns,
-                    values: vec![],
-                    default_values: false,
-                    select_source: Some(Box::new(sel)),
+                    source: InsertSource::Select(Box::new(sel)),
                     output,
                     output_into,
                 }));
-            } else {
-                return Err(DbError::Parse("expected SELECT in INSERT ... SELECT".into()));
             }
+        }
+    }
+
+    if let Some(exec_idx) = exec_idx_stripped {
+        if values_idx_stripped.is_none() || exec_idx < values_idx_stripped.unwrap() {
+            let head = after_into_stripped[..exec_idx].trim();
+            let (table_name, columns) = parse_table_and_columns(head)?;
+            let exec_sql = &after_into_stripped[exec_idx..];
+            let exec_stmt = crate::parser::parse_sql(exec_sql)?;
+            return Ok(Statement::Insert(InsertStmt {
+                table: parse_object_name(&table_name),
+                columns,
+                source: InsertSource::Exec(Box::new(exec_stmt)),
+                output,
+                output_into,
+            }));
         }
     }
 
     // INSERT ... VALUES
     let values_idx = values_idx_stripped
-        .ok_or_else(|| DbError::Parse("INSERT missing VALUES or SELECT".into()))?;
+        .ok_or_else(|| DbError::Parse("INSERT missing VALUES, SELECT or EXEC".into()))?;
 
     let head = after_into_stripped[..values_idx].trim();
     let values_part = after_into_stripped[values_idx + "VALUES".len()..].trim();
 
-    let (table_name, columns) = if let Some(open) = head.find('(') {
+    let (table_name, columns) = parse_table_and_columns(head)?;
+    let table = parse_object_name(&table_name);
+    let values = parse_values_groups(values_part)?;
+
+    Ok(Statement::Insert(InsertStmt {
+        table,
+        columns,
+        source: InsertSource::Values(values),
+        output,
+        output_into,
+    }))
+}
+
+fn parse_table_and_columns(head: &str) -> Result<(String, Option<Vec<String>>), DbError> {
+    if let Some(open) = head.find('(') {
         let close = head
             .rfind(')')
             .ok_or_else(|| DbError::Parse("missing ')' in INSERT columns".into()))?;
@@ -103,23 +114,10 @@ pub(crate) fn parse_insert(sql: &str) -> Result<Statement, DbError> {
             .split(',')
             .map(|c| c.trim().to_string())
             .collect::<Vec<_>>();
-        (table_name.to_string(), Some(cols))
+        Ok((table_name.to_string(), Some(cols)))
     } else {
-        (head.to_string(), None)
-    };
-
-    let table = parse_object_name(&table_name);
-    let values = parse_values_groups(values_part)?;
-
-    Ok(Statement::Insert(InsertStmt {
-        table,
-        columns,
-        values,
-        default_values: false,
-        select_source: None,
-        output,
-        output_into,
-    }))
+        Ok((head.to_string(), None))
+    }
 }
 
 pub(crate) fn parse_values_groups(input: &str) -> Result<Vec<Vec<Expr>>, DbError> {
