+++ crates/tsql_core/tests/dml_enhanced_test.rs
@@ -1,67 +1,73 @@
-use tsql_core::Engine;
+#[cfg(test)]
+mod dml_enhanced_tests {
+    use tsql_core::parser::parse_sql;
+    use tsql_core::ast::*;
 
-#[test]
-fn test_merge_with_conditions() {
-    let mut engine = Engine::new();
-    engine.exec("CREATE TABLE dbo.Target (Id INT PRIMARY KEY, Val NVARCHAR(50), Status INT)").unwrap();
-    engine.exec("CREATE TABLE dbo.Source (Id INT, Val NVARCHAR(50))").unwrap();
+    #[test]
+    fn test_update_top() {
+        let sql = "UPDATE TOP (10) t SET a = 1";
+        let stmt = parse_sql(sql).unwrap();
+        if let Statement::Update(u) = stmt {
+            assert!(u.top.is_some());
+            if let Expr::Integer(n) = u.top.unwrap().value {
+                assert_eq!(n, 10);
+            } else {
+                panic!("Expected integer top value");
+            }
+        } else {
+            panic!("Expected Update statement");
+        }
+    }
 
-    engine.exec("INSERT INTO dbo.Target (Id, Val, Status) VALUES (1, 'Old', 0), (2, 'Keep', 1)").unwrap();
-    engine.exec("INSERT INTO dbo.Source (Id, Val) VALUES (1, 'New'), (2, 'UpdateMe'), (3, 'BrandNew')").unwrap();
+    #[test]
+    fn test_delete_top() {
+        let sql = "DELETE TOP (5) FROM t WHERE a = 1";
+        let stmt = parse_sql(sql).unwrap();
+        if let Statement::Delete(d) = stmt {
+            assert!(d.top.is_some());
+            if let Expr::Integer(n) = d.top.unwrap().value {
+                assert_eq!(n, 5);
+            } else {
+                panic!("Expected integer top value");
+            }
+        } else {
+            panic!("Expected Delete statement");
+        }
+    }
 
-    // MERGE with conditions and multiple WHEN clauses
-    let sql = "
-        MERGE INTO dbo.Target AS T
-        USING dbo.Source AS S
-        ON T.Id = S.Id
-        WHEN MATCHED AND T.Status = 0 THEN
-            UPDATE SET T.Val = S.Val, T.Status = 1
-        WHEN MATCHED AND T.Status = 1 THEN
-            UPDATE SET T.Val = 'Updated'
-        WHEN NOT MATCHED THEN
-            INSERT (Id, Val, Status) VALUES (S.Id, S.Val, 2);
-    ";
+    #[test]
+    fn test_insert_exec() {
+        let sql = "INSERT INTO t (col1) EXEC my_proc @param1 = 1";
+        let stmt = parse_sql(sql).unwrap();
+        if let Statement::Insert(i) = stmt {
+            assert_eq!(i.table.name, "t");
+            assert_eq!(i.columns, Some(vec!["col1".to_string()]));
+            match i.source {
+                InsertSource::Exec(s) => {
+                    if let Statement::ExecProcedure(p) = *s {
+                        assert_eq!(p.name.name, "my_proc");
+                    } else {
+                        panic!("Expected ExecProcedure in InsertSource::Exec");
+                    }
+                }
+                _ => panic!("Expected InsertSource::Exec"),
+            }
+        } else {
+            panic!("Expected Insert statement");
+        }
+    }
 
-    engine.exec(sql).unwrap();
-
-    let result = engine.query("SELECT Id, Val, Status FROM dbo.Target ORDER BY Id").unwrap();
-    assert_eq!(result.rows.len(), 3);
-
-    // Id 1: matched, Status was 0 -> Updated to 'New', Status 1
-    assert_eq!(result.rows[0][1].to_string_value(), "New");
-    assert_eq!(result.rows[0][2].to_string_value(), "1");
-
-    // Id 2: matched, Status was 1 -> Updated to 'Updated', Status 1
-    assert_eq!(result.rows[1][1].to_string_value(), "Updated");
-    assert_eq!(result.rows[1][2].to_string_value(), "1");
-
-    // Id 3: not matched -> Inserted 'BrandNew', Status 2
-    assert_eq!(result.rows[2][0].to_string_value(), "3");
-    assert_eq!(result.rows[2][1].to_string_value(), "BrandNew");
-    assert_eq!(result.rows[2][2].to_string_value(), "2");
-}
-
-#[test]
-fn test_update_from_multiple_joins() {
-    let mut engine = Engine::new();
-    engine.exec("CREATE TABLE dbo.A (Id INT, Val INT)").unwrap();
-    engine.exec("CREATE TABLE dbo.B (Id INT, AId INT, Val INT)").unwrap();
-    engine.exec("CREATE TABLE dbo.C (Id INT, BId INT, Val INT)").unwrap();
-
-    engine.exec("INSERT INTO dbo.A (Id, Val) VALUES (1, 10)").unwrap();
-    engine.exec("INSERT INTO dbo.B (Id, AId, Val) VALUES (1, 1, 20)").unwrap();
-    engine.exec("INSERT INTO dbo.C (Id, BId, Val) VALUES (1, 1, 30)").unwrap();
-
-    let sql = "
-        UPDATE A
-        SET Val = A.Val + B.Val + C.Val
-        FROM dbo.A
-        INNER JOIN dbo.B ON A.Id = B.AId
-        INNER JOIN dbo.C ON B.Id = C.BId
-    ";
-
-    engine.exec(sql).unwrap();
-
-    let val = engine.query("SELECT Val FROM dbo.A WHERE Id = 1").unwrap().rows[0][0].to_string_value();
-    assert_eq!(val, "60"); // 10 + 20 + 30
+    #[test]
+    fn test_table_hints() {
+        let sql = "SELECT * FROM t WITH (NOLOCK, TABLOCK)";
+        let stmt = parse_sql(sql).unwrap();
+        if let Statement::Select(s) = stmt {
+            let tref = s.from.unwrap();
+            assert_eq!(tref.hints.len(), 2);
+            assert_eq!(tref.hints[0], "NOLOCK");
+            assert_eq!(tref.hints[1], "TABLOCK");
+        } else {
+            panic!("Expected Select statement");
+        }
+    }
 }
